pyproject.toml
[project]
name = "pdf-converter"
version = "0.1.0"
description = ""
authors = [
    {name = "al-ov73",email = "klever1987@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
